+++ src/main/java/com/lucendar/strm/common/strm/UriScheme.java
@@ -7,18 +7,24 @@
  *******************************************************************************/
 package com.lucendar.strm.common.strm;
 
+import com.lucendar.strm.common.StreamingApi;
+
 public enum UriScheme {
 
 
     HTTP("http"),
     HTTPS("https"),
     WS("ws"),
-    WSS("wss");
+    WSS("wss"),
+    RTSP("rtsp"),
+    RTMP("rtmp");
 
     public static final String HTTP_VALUE = "http";
     public static final String HTTPS_VALUE = "https";
     public static final String WS_VALUE = "ws";
     public static final String WSS_VALUE = "wss";
+    public static final String RTSP_VALUE = "rtsp";
+    public static final String RTMP_VALUE = "rtmp";
 
     private String str;
 
@@ -77,6 +83,8 @@ public enum UriScheme {
             case HTTPS_VALUE:
             case WS_VALUE:
             case WSS_VALUE:
+            case RTSP_VALUE:
+            case RTMP_VALUE:
                 return true;
 
             default:
@@ -98,8 +106,37 @@ public enum UriScheme {
             case WSS_VALUE:
                 return WSS;
 
+            case RTSP_VALUE:
+                return RTSP;
+
             default:
                 throw new IllegalArgumentException(String.format("Invalid UriScheme: `%s`.", uriScheme));
+        }
+    }
+
+    public static UriScheme defaultSchemeOf(int strmFormat, boolean sslEnabled) {
+        switch (strmFormat) {
+            case StreamingApi.STRM_FORMAT__HLS:
+                if (sslEnabled)
+                    return HTTPS;
+                else
+                    return HTTP;
+
+            case StreamingApi.STRM_FORMAT__RTSP:
+                return RTSP;
+
+            case StreamingApi.STRM_FORMAT__FLV:
+            case StreamingApi.STRM_FORMAT__RAW:
+                if (sslEnabled)
+                    return WSS;
+                else
+                    return WS;
+
+            case StreamingApi.STRM_FORMAT__RTMP:
+                return RTMP;
+
+            default:
+                throw new IllegalArgumentException(String.format("Stream format is not supported: `%d`.", strmFormat));
         }
     }
 }